fix(calendar): move utilities into separate file for Marko 4
