+++ src/components/ebay-calendar/index.marko
@@ -1,5 +1,5 @@
 // @ts-check
-import {localeOverride, toISO} from "./component"
+import { toISO } from "./date-utils"
 
 $ var numMonths = input.numMonths || 1;
 $ var monthDates = [...Array(numMonths)].map((_, i) => component.getMonthDate(state.offset + i));
